requirements.txt
bottle==0.12.13
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
oauthlib==2.1.0
PyDrive==1.3.1
requests==2.18.4
requests-oauthlib==1.0.0
